+++ src/add-ons/kernel/busses/virtio/virtio_mmio/VirtioMmioDevice.h
@@ -109,8 +109,8 @@ public:
 	VirtioMmioDevice();
 	status_t Init(phys_addr_t regs, size_t regsLen, int32 irq, int32 queueCnt);
 
-	status_t NegotiateFeatures(uint32 supported, uint32* negotiated, const char* (*get_feature_name)(uint32)) final;
-	status_t ClearFeature(uint32 feature) final;
+	status_t NegotiateFeatures(uint64 supported, uint64* negotiated, const char* (*get_feature_name)(uint64)) final;
+	status_t ClearFeature(uint64 feature) final;
 
 	status_t ReadDeviceConfig(uint8 offset, void* buffer, size_t bufferSize) final;
 	status_t WriteDeviceConfig(uint8 offset, const void* buffer, size_t bufferSize) final;
